Automate the tpen build.
* https://support.library.utoronto.ca/browse/NS-1836

Took out build.xml from .gitignore. The build definitions should be part
of the git repo to avoid "works on my machine" problems. Parameters can
be passed to ant via -Dwarfile=paleography.war for example.

The deploy target was left as is, but a package target has been added
for building the war. Ivy is now used to pull in dependencies, instead
of building against tomcat on the local machine. The version of tomcat
libs can be specified in ivy.xml. 7.0.52 is the version we are currently
running to serve tpen.

Added a publish target that sends the built war file to the nexus repo
defined in ivysettings-uoft.xml. The version number was taken from the
version.properties file, but is not dynamic.

There may need to be more work around parameterization of version
numbers and publishing destinations (e.g. snapshot or release, for
example).

The war file is now named based on the ivy module and nexus.user and
nexus.pass must be given as parameters in order to publish.