Update build process to be compatible with SLU
We now have access to SLU's git repository via sourceforge, which is
being mirrored on our gitlab instance for local modification.  We don't
want version.properties to be in our git archive so we're removing it
and it will later be generated by chef, as it is not used in the build
process.

* Update build.xml location
* Update ant basedir
* Remove version.properties